Separate out play function as its own module.
