+++ src/play.js
@@ -1,0 +1,27 @@
+/* globals contain */
+import { renderer } from "app";
+import { explorer } from "setup";
+function play() {
+    "use strict";
+
+    let collision = contain(explorer, {
+        x: 0,
+        y: 0,
+        width: renderer.view.width,
+        height: renderer.view.height
+    });
+    if (collision) {
+        if (collision.has("left") || collision.has("right")) {
+            explorer.vx = 0;
+        }
+        if (collision.has("up") || collision.has("down")) {
+            explorer.vy = 0;
+        }
+        return;
+    }
+
+    explorer.x += explorer.vx;
+    explorer.y += explorer.vy;
+}
+
+export { play };
